backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["phuyalbish <phuyalbish@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
pyjwt = "^2.8.0"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
